Add option for `helm-company` to use lower-cased company-prefix as Helm pattern
`helm-company-initialize-pattern-with-prefix` can now be either nil (do not
initialize pattern), t (initialize the prefix with the company prefix, exactly
as you typed it), and :downcase (same as t, but pass the prefix through
`downcase` first).

This stops Helm's smart case folding from getting "tricked" into being
case-sensitive if the prefix has uppercase characters in it.