FIX #20: Auto-complete without showing helm if there is only 1 candidate
This was already happening in the old code, inconsistently, depending on whether
or not helm-company was running `company-complete' itself or if it was relying
on a preexisting set of completion candidates.

I felt no need to preserve the old, inconsistent behaviour, where helm would pop
up to show even a single completion candidate.

Also, fixed a bug where helm-company would not have been shown if a
`company-point' DID exist, but the point was at 0. ;-)