FIX #21: Dynamic bindings were intercepting setq calls, preventing cleanup after completion
This bug only occurred when `company-idle-delay' was nil (or 0). meaning company
was not being run in the background automatically.

A dynamic-scope binding had been created for `company-candidates', which
variable is used by company to determine if completion is in progress. This
dynamic binding intercepted a "setq nil" call in the company cleanup code (in
`company-finish'). So, the "setq" call hit the dynamic bindings instead of the
global variable as would be required to actually clean them up.

So, the next time `company-post-command' was called after completion,
`company--perform' saw that there were `company-candidates' available, so there
must be a completion in progress! It tried to continue completion as though
company had already initiated completion. However, a completion was not
in-progress (there were just a few stale variables), so eventually the code
raised an error (it expected `company-point' to be non-nil).

HINT: Some of these functions ended up byte-compiled. It became much easier to
debug when I re-evaluated all of the defuns, which allowed for much more
ergonomic stepping-through the source code instead of several interspersed
bytecode functions.

FUTURE: I'm not sure why `helm-attrset', and `helm-attr', are being used in this
code base. Maybe for some kind of persistence? It all smells unnecessary to me,
but I don't have the time to boldly yank on that thread.

INTERESTING: While troubleshooting this, I had done the following:

  (put 'helm-company 'company-keep t) ;; After defun helm-company

and:

  (setq this-command 'helm-company) ;; inside defun helm-company-action-insert

This seems to have caused the *very next* call to `company-post-command'
to *not* fail, thanks to a certain conditional. The *next* time I ran any
command, *that* was when I saw the "wrong-type-argument number-or-marker-p"
error.

This gap made me suspect that either there was a bug in company (ultimately
false), or something wasn't being cleaned up correctly.

This was fun to debug.